refactor(benchmark.core): Move utils to the common level
