--- packages/benchmark.core/src/client/utils/group.ts
@@ -1,8 +1,0 @@
-import Group from '../group';
-
-/**
- * Returns true if something is a group.
- */
-export function isGroup(something: Object): something is Group {
-	return something instanceof Group;
-}
