--- packages/benchmark.core/src/client/utils/hook.ts
@@ -1,8 +1,0 @@
-import Hook from '../hook';
-
-/**
- * Returns true if something is a hook.
- */
-export function isHook(something: Object): something is Hook {
-	return something instanceof Hook;
-}
