+++ packages/benchmark.core/src/common/utils/hook.ts
@@ -1,6 +1,13 @@
 import Hook, { HookType } from '../../client/hook';
 
 /**
+ * Returns true if something is a hook.
+ */
+export function isHook(something: Object): something is Hook {
+	return something instanceof Hook;
+}
+
+/**
  * Returns hooks with specified types.
  */
 export function getHooksWithTypes(hooks: Hook[], types: HookType[]): Hook[] {
