--- packages/benchmark.core/src/worker/utils/index.ts
@@ -1,5 +1,0 @@
-import * as groupUtils from './group';
-import * as hookUtils from './hook';
-
-export const group = groupUtils;
-export const hook = hookUtils;
